vision/requirements.txt
opencv-contrib-python==4.5.3.56
Flask==1.0.*
numpy==1.20.0
Jinja2==2.10.*
markupsafe==2.0.1
itsdangerous==2.0.1
pyyaml==6.0
pyzmq==25.0.*
